# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog OpenApi Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/user-profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Profile",
                "operationId": "User Profile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change-Password",
                "operationId": "Change-Password",
                "parameters": [
                    {
                        "name": "old_password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/update-profile/{id}": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update Profile",
                "operationId": "Update Profile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/forget-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forget-Password",
                "operationId": "Forget-Password",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset-Password",
                "operationId": "Reset-Password",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation ",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
...
        "/api/update-blogpost/{id}": {
            "put": {
                "tags": [
                    "Blog Post"
                ],
                "summary": "update",
                "operationId": "update ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "APIKEY": {
                "type": "apiKey",
                "description": "Default Token :static",
                "name": "APIKEY",
                "in": "header"
            },
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}